Fix interval builder argument type

+++ app/src/main/java/at/plankt0n/wamediacopy/SettingsFragment.kt
@@ -294,7 +294,7 @@ class SettingsFragment : Fragment(),
         if (prefs.getLong(FileCopyWorker.PREF_LAST_COPY, 0L) == 0L) {
             prefs.edit().putBoolean(FileCopyWorker.PREF_REQUIRE_MANUAL_FIRST, true).apply()
         }
-        val period = minutes.coerceAtLeast(15)
+        val period = minutes.coerceAtLeast(15).toLong()
         val request = PeriodicWorkRequestBuilder<FileCopyWorker>(period, TimeUnit.MINUTES)
             .addTag(FileCopyWorker.TAG)
             .build()
@@ -322,7 +322,7 @@ class SettingsFragment : Fragment(),
 
         if (toggle.isChecked) {
             val minutes = prefs.getInt(FileCopyWorker.PREF_INTERVAL_MINUTES, 720)
-            val period = minutes.coerceAtLeast(15)
+            val period = minutes.coerceAtLeast(15).toLong()
             val requestPeriodic =
                 PeriodicWorkRequestBuilder<FileCopyWorker>(period, TimeUnit.MINUTES)
                     .addTag(FileCopyWorker.TAG)
